+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.4.4
 
 - Spell Volatility enchantment is now available for Weapons only
+- Spell Volatility and Amplify Spell enchantments are now mutually exclusive
 - Enchantment descriptions now have a trailing dot, to be consistent with vanilla enchantments
 - Add effect descriptions support
 - Add innate spell power bonus